+++ src/hooks/index.js
@@ -10,6 +10,5 @@ export * from './useFetchData';
 export * from './usePrevious';
 export * from './useTransfer';
 export * from './useConfig';
-export * from './useEventListener';
 export * from './useTransferProgress';
 export * from './useMaxAmount';
